requirements.txt
fastapi
uvicorn
pandas
scikit-learn
joblib
google-generativeai
pytest
python-dotenv